extract clerk key handling into AnonLayout
